+++ src/GreenSale.WebApi/Controllers/Admin/AdminBaseController.cs
@@ -1,8 +1,9 @@
 ﻿using GreenSale.WebApi.Controllers.Common;
+using GreenSale.WebApi.Controllers.Heat;
 using Microsoft.AspNetCore.Authorization;
 
 namespace GreenSale.WebApi.Controllers.Admin;
 
 [Authorize(Roles = "Admin")]
-public class AdminBaseController : BaseController
+public class AdminBaseController : SuperAdminBaseController
 { }
