+++ src/GreenSale.WebApi/Controllers/Admin/Categories/AdminCategoriesController.cs
@@ -1,56 +1,53 @@
-﻿using GreenSale.Application.Utils;
-using GreenSale.Persistence.Dtos.CategoryDtos;
+﻿using GreenSale.Persistence.Dtos.CategoryDtos;
 using GreenSale.Persistence.Validators.Categories;
 using GreenSale.Service.Interfaces.Categories;
 using Microsoft.AspNetCore.Mvc;
 
-namespace GreenSale.WebApi.Controllers.Admin.Categories
+namespace GreenSale.WebApi.Controllers.Admin.Categories;
+
+[Route("api/admin/categories")]
+[ApiController]
+public class AdminCategoriesController : AdminBaseController
 {
-    [Route("api/admin/categories")]
-    [ApiController]
-    public class AdminCategoriesController : AdminBaseController
+    private ICategoryService _service;
+
+    public AdminCategoriesController(ICategoryService service)
     {
-        private ICategoryService _service;
-        private readonly int maxPage = 30;
+        _service = service;
+    }
 
-        public AdminCategoriesController(ICategoryService service)
+
+    [HttpGet("count")]
+    public async Task<IActionResult> CountAsync()
+        => Ok(await _service.CountAsync());
+
+    [HttpDelete("{categoryId}")]
+    public async Task<IActionResult> DeleteAsync(long categoryId)
+        => Ok(await _service.DeleteAsync(categoryId));
+
+    [HttpPost]
+    public async Task<IActionResult> CreateAsync([FromForm] CategoryCreateDto dto)
+    {
+        CategoryCreateValidator validations = new CategoryCreateValidator();
+        var result = validations.Validate(dto);
+
+        if (result.IsValid)
         {
-            _service = service;
+            return Ok(await _service.CreateAsync(dto));
         }
 
+        return BadRequest(result.Errors);
+    }
 
-        [HttpGet("count")]
-        public async Task<IActionResult> CountAsync()
-            => Ok(await _service.CountAsync());
-
-        [HttpDelete("{categoryId}")]
-        public async Task<IActionResult> DeleteAsync(long categoryId)
-            => Ok(await _service.DeleteAsync(categoryId));
-
-        [HttpPost]
-        public async Task<IActionResult> CreateAsync([FromForm] CategoryCreateDto dto)
+    [HttpPut("{categoryId}")]
+    public async Task<IActionResult> UpdateAsync(long categoryId, [FromForm] CategoryCreateDto dto)
+    {
+        CategoryCreateValidator update = new CategoryCreateValidator();
+        var validationResult = update.Validate(dto);
+        if (validationResult.IsValid)
         {
-            CategoryCreateValidator validations = new CategoryCreateValidator();
-            var result = validations.Validate(dto);
-
-            if (result.IsValid)
-            {
-                return Ok(await _service.CreateAsync(dto));
-            }
-
-            return BadRequest(result.Errors);
+            return Ok(await _service.UpdateAsync(categoryId, dto));
         }
-
-        [HttpPatch("{categoryId}")]
-        public async Task<IActionResult> UpdateAsync(long categoryId, [FromForm] CategoryCreateDto dto)
-        {
-            CategoryCreateValidator update = new CategoryCreateValidator();
-            var validationResult = update.Validate(dto);
-            if (validationResult.IsValid)
-            {
-                return Ok(await _service.UpdateAsync(categoryId, dto));
-            }
-            return BadRequest(validationResult.Errors);
-        }
+        return BadRequest(validationResult.Errors);
     }
 }
