+++ src/GreenSale.WebApi/Controllers/Admin/Roles/AdminRolesController.cs
@@ -10,38 +10,12 @@ namespace GreenSale.WebApi.Controllers.Admin.Roles
     [ApiController]
     public class AdminRolesController : AdminBaseController
     {
-        public int maxPage = 30;
         private readonly IRoleService _service;
 
         public AdminRolesController(IRoleService roleService)
         {
             this._service = roleService;
         }
-
-        [HttpGet]
-        public async Task<IActionResult> GetAllAsync([FromQuery] int page = 1)
-            => Ok(await _service.GetAllAsync(new PaginationParams(page, maxPage)));
-
-        [HttpGet("{roleId}")]
-        public async Task<IActionResult> GetByIdAsync(long roleId)
-            => Ok(await _service.GetByIdAsync(roleId));
-
-        [HttpGet("count")]
-        public async Task<IActionResult> CountAsync()
-            => Ok(await _service.CountAsync());
-
-        [HttpPost]
-        public async Task<IActionResult> CreateAsync([FromForm] RoleCreatDto dto)
-        {
-            RoleValidator validationRules = new RoleValidator();
-            var result = validationRules.Validate(dto);
-
-            if (result.IsValid)
-                return Ok(await _service.CreateAsync(dto));
-
-            return BadRequest(result.Errors);
-        }
-
         [HttpPut("{roleId}")]
         public async Task<IActionResult> UpdateAsync(long roleId, RoleCreatDto dto)
         {
@@ -53,9 +27,5 @@ namespace GreenSale.WebApi.Controllers.Admin.Roles
 
             return BadRequest(result.Errors);
         }
-
-        [HttpDelete("{roleId}")]
-        public async Task<IActionResult> DeleteAsync(long roleId)
-            => Ok(await _service.DeleteAsync(roleId));
     }
 }
