+++ src/GreenSale.WebApi/Controllers/Heat/Roles/UserRoleController.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+
+namespace GreenSale.WebApi.Controllers.Heat.Roles
+{
+    [Route("api/heat/superadmin/user/roles")]
+    [ApiController]
+    public class UserRoleController : SuperAdminBaseController
+    {
+        public UserRoleController()
+        {
+            
+        }
+    }
+}
